\name{cpppkg-package}
\alias{cpppkg-package}
\alias{cpppkg}
\docType{package}
\title{
\packageTitle{cpppkg}
}
\description{
\packageDescription{cpppkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cpppkg}
\packageIndices{cpppkg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{cpppkg}

Maintainer: \packageMaintainer{cpppkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
